move CallTraceReader to its own namespace
