+++ src/Command/Inspector/GetTraceCommand.php
@@ -28,9 +28,9 @@ use Reli\Lib\Elf\Process\ProcessSymbolReaderException;
 use Reli\Lib\Elf\Tls\TlsFinderException;
 use Reli\Lib\PhpProcessReader\PhpGlobalsFinder;
 use Reli\Lib\PhpProcessReader\PhpVersionDetector;
-use Reli\Lib\PhpProcessReader\TraceCache;
 use Reli\Lib\Process\MemoryReader\MemoryReaderException;
-use Reli\Lib\PhpProcessReader\PhpMemoryReader\CallTraceReader;
+use Reli\Lib\PhpProcessReader\CallTraceReader\CallTraceReader;
+use Reli\Lib\PhpProcessReader\CallTraceReader\TraceCache;
 use Reli\Lib\Process\ProcessStopper\ProcessStopper;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
