+++ src/Inspector/Daemon/Reader/Protocol/Message/TraceMessage.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Reli\Inspector\Daemon\Reader\Protocol\Message;
 
-use Reli\Lib\PhpProcessReader\CallTrace;
+use Reli\Lib\PhpProcessReader\CallTraceReader\CallTrace;
 
 final class TraceMessage
 {
