+++ src/Inspector/Output/TraceFormatter/CallTraceFormatter.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Reli\Inspector\Output\TraceFormatter;
 
-use Reli\Lib\PhpProcessReader\CallTrace;
+use Reli\Lib\PhpProcessReader\CallTraceReader\CallTrace;
 
 interface CallTraceFormatter
 {
