+++ src/Lib/PhpProcessReader/CallTraceReader/TraceCache.php
@@ -11,8 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Reli\Lib\PhpProcessReader;
+namespace Reli\Lib\PhpProcessReader\CallTraceReader;
 
+use Reli\Lib\Process\Pointer\Dereferencable as T;
 use Reli\Lib\Process\Pointer\Dereferencer;
 use Reli\Lib\Process\Pointer\Pointer;
