+++ tests/Inspector/Output/TraceFormatter/Templated/TemplatedCallTraceFormatterTest.php
@@ -13,8 +13,8 @@ declare(strict_types=1);
 
 namespace Reli\Inspector\Output\TraceFormatter\Templated;
 
-use Reli\Lib\PhpProcessReader\CallFrame;
-use Reli\Lib\PhpProcessReader\CallTrace;
+use Reli\Lib\PhpProcessReader\CallTraceReader\CallFrame;
+use Reli\Lib\PhpProcessReader\CallTraceReader\CallTrace;
 use PHPUnit\Framework\TestCase;
 
 class TemplatedCallTraceFormatterTest extends TestCase
