+++ tests/Lib/PhpProcessReader/CallTraceReader/CallTraceReaderTest.php
@@ -11,10 +11,10 @@
 
 declare(strict_types=1);
 
-namespace Reli\Lib\PhpProcessReader\PhpMemoryReader;
+namespace Reli\Lib\PhpProcessReader\CallTraceReader;
 
+use PHPUnit\Framework\TestCase;
 use Reli\Inspector\Settings\TargetPhpSettings\TargetPhpSettings;
-use Reli\Inspector\Settings\TargetProcessSettings\TargetProcessSettings;
 use Reli\Lib\ByteStream\IntegerByteSequence\LittleEndianReader;
 use Reli\Lib\Elf\Parser\Elf64Parser;
 use Reli\Lib\Elf\Process\PerBinarySymbolCacheRetriever;
@@ -22,22 +22,13 @@ use Reli\Lib\Elf\Process\ProcessModuleSymbolReaderCreator;
 use Reli\Lib\Elf\SymbolResolver\Elf64SymbolResolverCreator;
 use Reli\Lib\File\CatFileReader;
 use Reli\Lib\PhpInternals\Opcodes\OpcodeFactory;
-use Reli\Lib\PhpInternals\Types\Zend\ZendCastedTypeProvider;
-use Reli\Lib\PhpInternals\Types\Zend\ZendExecuteData;
-use Reli\Lib\PhpInternals\Types\Zend\ZendExecutorGlobals;
-use Reli\Lib\PhpInternals\Types\Zend\ZendFunction;
-use Reli\Lib\PhpInternals\Types\Zend\ZendString;
 use Reli\Lib\PhpInternals\ZendTypeReader;
 use Reli\Lib\PhpInternals\ZendTypeReaderCreator;
-use Reli\Lib\PhpProcessReader\TraceCache;
-use Reli\Lib\Process\MemoryMap\ProcessMemoryMapCreator;
-use Reli\Lib\Process\MemoryReader\MemoryReader;
 use Reli\Lib\PhpProcessReader\PhpGlobalsFinder;
 use Reli\Lib\PhpProcessReader\PhpSymbolReaderCreator;
-use Reli\Lib\Process\Pointer\Pointer;
-use Reli\Lib\Process\Pointer\RemoteProcessDereferencer;
+use Reli\Lib\Process\MemoryMap\ProcessMemoryMapCreator;
+use Reli\Lib\Process\MemoryReader\MemoryReader;
 use Reli\Lib\Process\ProcessSpecifier;
-use PHPUnit\Framework\TestCase;
 
 class CallTraceReaderTest extends TestCase
 {
